[FEAT] 애플로그인, 카카오 로그인, 애플 로그인으로 인한 Sign변경

+++ Vodam/Feature/Main/Setting/SettingsFeature.swift
@@ -9,6 +9,7 @@ import ComposableArchitecture
 import Foundation
 import PhotosUI
 import SwiftUI
+import UIKit
 
 @Reducer
 struct SettingsFeature {
@@ -18,6 +19,9 @@ struct SettingsFeature {
         var user: User?
 
         @Presents var alert: AlertState<Action.Alert>?
+
+        var lastDeletedOwnerId: String? = nil
+        var isShowingAppleDisconnectGuide: Bool = false
     }
 
     enum Action: Equatable {
@@ -30,6 +34,10 @@ struct SettingsFeature {
         case logoutFinished(Bool)
         case deleteAccountFinished(Bool)
 
+        case appleDisconnectGuideOpenSettingsButtonTapped
+        case appleDisconnectGuideCompletedButtonTapped
+        case appleDisconnectGuideDismissed
+
         case profileImagePicked(Data)
         case photoPickerItemChanged(PhotosPickerItem?)
 
@@ -39,7 +47,8 @@ struct SettingsFeature {
 
         enum Delegate: Equatable {
             case userUpdated(User)
-            case accountCleared(Bool)
+            case loggedOut(Bool)
+            case accountDeleted(Bool)
         }
 
         enum Alert: Equatable {
@@ -51,6 +60,7 @@ struct SettingsFeature {
     @Dependency(\.googleAuthClient) var googleAuthClient
     @Dependency(\.kakaoAuthClient) var kakaoAuthClient
     @Dependency(\.appleAuthClient) var appleAuthClient
+    @Dependency(\.firebaseClient) var firebaseClient
 
     var body: some Reducer<State, Action> {
         Reduce { state, action in
@@ -117,10 +127,10 @@ struct SettingsFeature {
             case .logoutFinished(let isSuccess):
                 if isSuccess {
                     state.user = nil
-                    return .send(.delegate(.accountCleared(true)))
+                    return .send(.delegate(.loggedOut(true)))
                 } else {
                     print("로그아웃 실패")
-                    return .send(.delegate(.accountCleared(false)))
+                    return .send(.delegate(.loggedOut(false)))
                 }
 
             case .deleteAccountTapped:
@@ -146,11 +156,21 @@ struct SettingsFeature {
                     return .send(.deleteAccountFinished(false))
                 }
 
+                if user.provider == .apple {
+                    state.isShowingAppleDisconnectGuide = true
+                    return .none
+                }
+
+                let ownerId = user.ownerId
+
                 switch user.provider {
                 case .kakao:
                     return .run { send in
                         do {
                             try await kakaoAuthClient.deleteAccount()
+                            try await firebaseClient.deleteAllForUser(
+                                user.ownerId
+                            )
                             await send(.deleteAccountFinished(true))
                         } catch {
                             print("카카오 회원 탈퇴 실패: \(error)")
@@ -162,6 +182,9 @@ struct SettingsFeature {
                     return .run { send in
                         do {
                             try await googleAuthClient.disconnect()
+                            try await firebaseClient.deleteAllForUser(
+                                user.ownerId
+                            )
                             await send(.deleteAccountFinished(true))
                         } catch {
                             print("구글 계정 연결 해제 실패: \(error)")
@@ -170,24 +193,49 @@ struct SettingsFeature {
                     }
 
                 case .apple:
-                    return .run { [appleAuthClient] send in
-                        do {
-                            try await appleAuthClient.deleteAccount()
-                            await send(.deleteAccountFinished(true))
-                        } catch {
-                            print("애플 회원 탈퇴 실패: \(error)")
-                            await send(.deleteAccountFinished(false))
-                        }
+                    return .none
+                }
+
+            case .appleDisconnectGuideOpenSettingsButtonTapped:
+                return .run { _ in
+                    if let url = URL(
+                        string: UIApplication.openSettingsURLString
+                    ) {
+                        UIApplication.shared.open(url)
                     }
                 }
+
+            case .appleDisconnectGuideCompletedButtonTapped:
+                state.isShowingAppleDisconnectGuide = false
+
+                guard let user = state.user else {
+                    return .send(.deleteAccountFinished(false))
+                }
+                let ownerId = user.ownerId
+
+                return .run { send in
+                    do {
+                        try await firebaseClient.deleteAllForUser(ownerId)
+                        await send(.deleteAccountFinished(true))
+                    } catch {
+                        print("애플 회원 탈퇴 실패: \(error)")
+                        await send(.deleteAccountFinished(false))
+                    }
+                }
+
+            case .appleDisconnectGuideDismissed:
+                state.isShowingAppleDisconnectGuide = false
+                return .none
 
             case .deleteAccountFinished(let isSuccess):
                 if isSuccess {
                     state.user = nil
-                    return .send(.delegate(.accountCleared(true)))
+                    state.lastDeletedOwnerId = nil
+                    return .send(.delegate(.accountDeleted(true)))
                 } else {
                     print("회원 탈퇴 실패")
-                    return .send(.delegate(.accountCleared(false)))
+                    state.lastDeletedOwnerId = nil
+                    return .send(.delegate(.accountDeleted(false)))
                 }
 
             case .photoPickerItemChanged(let item):
